gaming-api/requirements.txt
fastapi
uvicorn
python-dotenv
requests
sqlalchemy
pydantic
psycopg2-binary
